src/main/java/pro/sky/questiongencw2/service/JavaQuestionService.java
package pro.sky.questiongencw2.service;

import pro.sky.questiongencw2.domain.Question;
import pro.sky.questiongencw2.exceptions.NotFoundException;
import org.springframework.stereotype.Service;

import java.util.*;

@Service
public class JavaQuestionService implements QuestionService {

    private final Set<Question> questions;

    public JavaQuestionService() {
        this.questions = new HashSet<>();
    }

    @Override
    public Question add(String question, String answer) {
        Question quest = new Question(question, answer);
        questions.add(quest);
        return quest;
    }

    @Override
    public Question add(Question question) {
        questions.add(question);
        return question;
    }

    @Override
    public Question remove(Question question) {
        if (!questions.contains(question)) {
            throw new NotFoundException();
        }
        questions.remove(question);
        return question;
    }

    @Override
    public Collection<Question> getAll() {
        return Collections.unmodifiableSet(questions);
    }

    @Override
    public Question getRandomQuestion() {
        int random = new Random().nextInt(questions.size());
        int i = 0;
        for (Question question : questions) {
            if (i == random) {
                return question;
            }
            i++;
        }
        return null;
    }
}
